requirements.txt
asgiref
certifi
charset-normalizer
django>=5.2,<6
idna
python-decouple
requests
sqlparse
urllib3
pillow
djangorestframework
markdown
django-filter
drf-yasg
drf-yasg[validation]
djangorestframework-simplejwt
django-mail-templated
djoser
black
flake8
django-simple-captcha